Added Sync Loop
Sync loop will sync the start and end of loops to the start of the
next iteration of the first loop. Uses the looper.offset to keep
the loops in sync when there is an offset on the first loop, so added
the ability to set the offset in looper as well, not just
the LooperBackend.